requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
backoff==2.2.1
beautifulsoup4==4.14.3
browser-use==0.12.5
browser-use-sdk==2.0.15
bubus==1.5.6
cbor2==5.8.0
cdp-use==1.4.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
constantly==23.10.4
cryptography==46.0.5
Cython==3.2.4
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
dotenv==0.9.9
executing==2.2.1
fastuuid==0.14.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.65.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==1.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
hyperlink==21.0.0
idna==3.11
importlib_metadata==9.0.0
Incremental==24.11.0
inquirerpy==0.3.4
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lance-namespace==0.5.2
lance-namespace-urllib3-client==0.5.2
lancedb==0.29.2
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-openai==1.1.9
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
litellm==1.82.2
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
oauthlib==3.3.1
ollama==0.6.1
openai==2.16.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pfzy==0.3.4
pillow==12.1.1
playwright==1.58.0
portalocker==3.2.0
posthog==7.7.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.10.1
pyobjc-core; sys_platform == "darwin"
pyobjc-framework-Cocoa; sys_platform == "darwin"
pyOpenSSL==26.0.0
pyotp==2.9.0
pyparsing==3.3.2
pypdf==6.9.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
screeninfo==0.8.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
sse-starlette==3.3.3
stack-data==0.6.3
starlette==1.0.0
tantivy==0.25.1
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
uuid7==0.1.0
uuid_utils==0.14.1
uvicorn==0.42.0
wcwidth==0.6.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zope.interface==8.2
zstandard==0.25.0
redis>=5.0.0
